refactoring row adapter and row wrapper classes
