+++ common/core/src/main/java/zingg/common/core/documenter/RowWrapper.java
@@ -1,4 +1,4 @@
-package zingg.common.core.util;
+package zingg.common.core.documenter;
 
 import freemarker.template.Version;
 import freemarker.template.DefaultObjectWrapper;
